FEAT: 코인 증가 기능

+++ backend/src/main/java/org/ftclub/cabinet/user/domain/User.java
@@ -147,7 +147,7 @@ public class User {
 		return blackholedAt != null && blackholedAt.isBefore(LocalDateTime.now());
 	}
 
-	public void addCoin(int reward) {
+	public void addCoin(Long reward) {
 		this.coin += reward;
 	}
 }
